Authenticate is invoked on client initialization
There's little value in a client that is not already authenticated.
Making the authenticate method private and called on the initialization
of a NewClient gives the user a better experience. There's no need to
create a NewClient and then try to use it, only to see that it can not
be authenticated. Providing that feedback at initialization will save
the user that time.

The testing strategy is expanded to create a mock for the authorization
portion. The builder library is leverage to create accessToken structs
which can be marshaled into the responses that we wish the mock api to
return.